+++ src/components/PluginPresets/PresetDropdown/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../../../styles/template-styles/Modal.module.scss';
+import styles from '../../../styles/template-styles/Plugin.module.scss';
 import stylesPresets from '../../../styles/template-styles/Presets.module.scss';
 import { IPresetDropdownProps } from '../../../utils/Interfaces/PluginPresets/Dropdown.interface';
 import { PresetHandleAction } from '../../../utils/constants';
